don't hard-code settings directory
Before, the directory qhub would load settings from was set at compile time
through a rather ugly CFLAGS hack. Now, it tries the first of the following
locations that is specified:
* the directory passed on the command line using --config-dir
* the directory in the QHUB_CONFIG environment variable
* the current directory

Since the idea is for it to be started mainly by init scripts, which can easily
set the command line correctly, this shouldn't cause any usability issues and
should increase flexibility.